+++ bootloader/debug.h
@@ -1,0 +1,12 @@
+#ifndef BOOTLOADER_DEBUG_H
+#define BOOTLOADER_DEBUG_H
+
+#ifdef DEBUG
+#define debug_printf(fmt, ...)   fprintf(stderr, fmt, ##__VA_ARGS__)
+#define TAR_DEBUG_OPTIONS		(TAR_VERBOSE)
+#else
+#define debug_printf(fmt, ...)
+#define TAR_DEBUG_OPTIONS		0
+#endif
+
+#endif /* BOOTLOADER_DEBUG_H */