+++ staticx/archive.py
@@ -2,14 +2,35 @@ import tarfile
 import logging
 from os.path import basename, islink
 
+try:
+    # Python >= 3.3
+    import lzma
+    lzma.open       # pyliblzma doesn't have lzma.open()
+except (ImportError, AttributeError):
+    # Python < 3.3
+    from backports import lzma
+
 from .utils import get_symlink_target
 from .constants import *
 from .errors import *
 
-
 class SxArchive(object):
     def __init__(self, fileobj, mode):
-        self.tar = tarfile.open(fileobj=fileobj, mode=mode)
+        self.xzf = lzma.open(
+            filename = fileobj,
+            mode = mode,
+            format = lzma.FORMAT_XZ,
+
+            # Use CRC32 instead of CRC64 (FORMAT_XZ default)
+            # Otherwise, enable XZ_USE_CRC64 in libxz/xz_config.h
+            check = lzma.CHECK_CRC32,
+
+            filters = [
+                dict(id=lzma.FILTER_X86),
+                dict(id=lzma.FILTER_LZMA2)
+            ],
+        )
+        self.tar = tarfile.open(fileobj=self.xzf, mode=mode)
         self._added_libs = []
 
     def __enter__(self):
@@ -17,6 +38,8 @@ class SxArchive(object):
 
     def __exit__(self, *excinfo):
         self.tar.close()
+        self.xzf.close()
+
 
     @property
     def libraries(self):